Extract ref-kind call-site fixture
