+++ src/ILInspector.Decompiler.Tests/CfgSampleClass.cs
@@ -5374,19 +5374,6 @@ public class CfgSampleClass
     }
 }
 
-public sealed class RefKindBox<T>
-{
-    T _value = default!;
-
-    public bool TryGet(out T value)
-    {
-        value = _value;
-        return true;
-    }
-
-    public void Put(in T value) => _value = value;
-}
-
 public interface IJoinShape
 {
     string Shape();
