+++ src/ILInspector.Decompiler.Tests/RefKindCallSiteSamples.cs
@@ -1,0 +1,14 @@
+namespace ILInspector.Decompiler.Tests;
+
+public sealed class RefKindBox<T>
+{
+    T _value = default!;
+
+    public bool TryGet(out T value)
+    {
+        value = _value;
+        return true;
+    }
+
+    public void Put(in T value) => _value = value;
+}
